Remove tenant database code
