+++ cmd/authgear/elasticsearch/wire.go
@@ -8,7 +8,7 @@ import (
 	"github.com/google/wire"
 
 	"github.com/authgear/authgear-server/pkg/lib/config"
-	tenantdb "github.com/authgear/authgear-server/pkg/lib/infra/db/tenant"
+	"github.com/authgear/authgear-server/pkg/lib/infra/db"
 )
 
 func NewAppLister(
@@ -20,7 +20,7 @@ func NewAppLister(
 
 func NewReindexer(
 	ctx context.Context,
-	pool *tenantdb.Pool,
+	pool *db.Pool,
 	databaseCredentials *config.DatabaseCredentials,
 	appID config.AppID,
 ) *Reindexer {
