+++ pkg/lib/event/service.go
@@ -7,7 +7,7 @@ import (
 	"github.com/authgear/authgear-server/pkg/api/event"
 	"github.com/authgear/authgear-server/pkg/api/model"
 	"github.com/authgear/authgear-server/pkg/lib/config"
-	tenantdb "github.com/authgear/authgear-server/pkg/lib/infra/db/tenant"
+	"github.com/authgear/authgear-server/pkg/lib/infra/db"
 	"github.com/authgear/authgear-server/pkg/lib/session"
 	"github.com/authgear/authgear-server/pkg/util/clock"
 	"github.com/authgear/authgear-server/pkg/util/intl"
@@ -21,7 +21,7 @@ type UserService interface {
 }
 
 type Database interface {
-	UseHook(hook tenantdb.TransactionHook)
+	UseHook(hook db.TransactionHook)
 }
 
 type Sink interface {
