--- pkg/lib/infra/db/tenant/hook.go
@@ -1,6 +1,0 @@
-package tenant
-
-type TransactionHook interface {
-	WillCommitTx() error
-	DidCommitTx()
-}
